Move StakeStore from `persistence` to `signer`
