+++ internal/mithril-persistence/src/store/mod.rs
@@ -5,5 +5,5 @@ pub mod adapter;
 mod stake_store;
 mod store_pruner;
 
-pub use stake_store::{StakeStore, StakeStorer};
+pub use stake_store::StakeStorer;
 pub use store_pruner::StorePruner;
