+++ mithril-signer/src/store/mod.rs
@@ -2,6 +2,8 @@
 
 mod mktree_store_sqlite;
 mod protocol_initializer_store;
+mod stake_store;
 
 pub use mktree_store_sqlite::*;
 pub use protocol_initializer_store::*;
+pub use stake_store::*;
